Remove check for IBM cable card
Removing the check for IBM cable card because this will be
verified in VPD and there is no need to have an additional
check inside PLDM.
This commit also includes the check for presence property
before PDRs are added for adapters and ports. PDRs should
not be added for empty slots.

Tested by : performing CM on IBM cable cards and industry
standard adapters

Signed-off-by: Varsha Kaverappa <vkaverap@in.ibm.com>