oem-ibm: Improving SBE Dump
Earlier, we were using the getSubTree API, which returns a pair of
object paths and a mapper service map. From this, we extracted the
object path and stored it for future retrieval of the entity ID from
that path. We are now replacing this API with the existing
getSubTreePaths API, as it returns only the object paths, which we use
to extract entity IDs. Additionally, the function to extract the entity
IDs now utilizes regex for a more dynamic approach instead of using
hardcoded character position indices to extract the entity ID.

Tested: Verified the PDRs for populating correct entity instance IDs.

Signed-off-by: Riya Dixit <riyadixitagra@gmail.com>