oem: Check presence of the fru for topology information
Adding a presence check of the fru before populating the
topology information.

The Slot had JBOF card connected earlier and it was removed and cable
card was added in the same slot. So the Dbus inventory under the PCIE
card contain all the connectors underneath JBOF card as well as cable
card and "Present" property provides the information about which
connectors are active at the moment.
PLDM also need to check "Present" property to fetch the right
object path to create the proper association.

Signed-off-by: Archana Kakani <archana.kakani@ibm.com>