oem: Check presence of the fru for topology information
Adding a presence check of the fru before populating the
topology information.

If the JBOF card is replaced by cable card in the slot, the
Dbus inventory under the PCIe card contain all the connectors
underneath JBOF card as well as cable card and "Present" property
provides the information about which connectors are active at the moment.
PLDM need to check "Present" property to fetch the right object path to
create the proper association.

Signed-off-by: Archana Kakani <archana.kakani@ibm.com>