oem-ibm: Fixes the issue in resource dump entry
When a resource dump is created with an empty resource selector
from GUI, a system dump is generated and this was entering our
resource dump path and was creating wrong entry name.

This is seen in 1110 because of the recent changes in the
naming format of the dumps.
Resource dump starts with 'B' and System dump starts with 'A'.

Tested with the fix and it worked good.

Change-Id: Iac7d1455f568e928132c80abb562978184109294
Signed-off-by: Pavithra Barithaya <pavithrabarithaya07@gmail.com>