host-bmc: Fix in the panel flow for Reset Reload (#690)
The op-panel function enabling will be sent down from PHYP with a
modify event after the PDR exchange in the normal poweron path.
But after a Reset Reload the function bitmaps will be sent down
from PHYP with a Added event. To trigger the panel app even during
the added event (reset reload case) this change is made.

Tested: Triggered the functions using the paneltool and the
functions are persisted post reset reload also.

Signed-off-by: Pavithra Barithaya <pavithrabarithaya07@gmail.com>