oem-ibm: Invalid pel fix (#691)
Phyp copies the error log data to the XDMA memory and then it sends the
`WriteFileByTypeFromMemory` to PLDM.  From PLDM side, setting the struct
AspeedXdmaOp instructs DMA controller to copy the data to bmc side.
Because of slow data copy, we have partial data in the BMC XDMA buffer.
As we have mmapped the VGA memory to the length of the Pel, we fetch the
exact length which contains partial data and remaining bytes as garbage.

To fix the issue, added a check to make sure that if the DMA data copy
is completed successfuly before copying the pel data to temporary file.

Also added a fix to avoid freeing instance IDs acquired by other apps.
And also sending the panel bitmap request only once to panel app

Fixes: https://jazz07.rchland.ibm.com:13443/jazz/web/projects/CSSD#action=com.ibm.team.workitem.viewWorkItem&id=678010

Signed-off-by: Archana Kakani <archana.kakani@ibm.com>