oem-ibm: NVRAM service file changes.
The NVRAM file size could be wrong in manufactuing itself or
somewhere in between the process and it becomes difficult to check
where could the file size go wrong.
Earlier PLDM service would not create the NVRAM file if it already
exists. But the file could have the wrong file size which could
lead to issues.
There was a need to check the NVRAM file size.

This change in PLDM NVRAM service files handles below conditions
1. If the file /var/lib/phosphor-software-manager/hostfw/nvram/PHYP-NVRAM
   exists and if it has a bad size we have to remove it and recreate it.
2. if /var/lib/phosphor-software-manager/hostfw/nvram/PHYP-NVRAM file does not exist
   then
   check if /var/lib/pldm/PHYP-NVRAM exists
   if so then move /var/lib/pldm/PHYP-NVRAM to /var/lib/phosphor-software-manager/hostfw/nvram/PHYP-NVRAM
   if not...or if the file exists with bad size then
   create /var/lib/phosphor-software-manager/hostfw/nvram/PHYP-NVRAM using truncate

Signed-off-by: Pavithra Barithaya <pavithrabarithaya07@gmail.com>