pldm: Core Count and DIMM Persistency (#701)
In the current code we use Set-Property to set the Core count.
This value is reverted to 0 post reboot as inventory manager will
only persist the data if it is set through Notify.

Changes added to persist the CPU core count and DIMM Operational
status across reboots when host is powered off.

Fixes: STGDefect 709813

Signed-off-by: Archana Kakani <archana.kakani@ibm.com>